requirements.txt
fastapi
uvicorn
textblob
google-generativeai
SpeechRecognition
pydub
python-multipart
gtts